net: shell: Add websocket command
Move "websocket" command to a separate file.
