+++ subsys/net/lib/shell/CMakeLists.txt
@@ -30,6 +30,7 @@ zephyr_library_sources(tcp.c)
 zephyr_library_sources(udp.c)
 zephyr_library_sources(virtual.c)
 zephyr_library_sources(vlan.c)
+zephyr_library_sources(websocket.c)
 
 zephyr_library_include_directories(
   ${ZEPHYR_BASE}/kernel/include
